latin_russian/requirements.txt
attrs==21.4.0
black==22.1.0
click==8.0.4
flake8==4.0.1
iniconfig==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pydantic==1.9.0
pyflakes==2.4.0
pyparsing==3.0.7
pytest==7.0.1
PyYAML==6.0
tomli==2.0.1
typing_extensions==4.1.1
